docker102: fix multi stage demo and add base-chroot (#25)
The example was throwing errors because we weren't copying all the
required libraries for `jq`.

I was also finding it hard to remember all the steps to check the
dependencies of `jq` with `apk info -L` and `ldd`, so I've explicitly
added them to the demo.

Also tweak the layer demo slightly to show diffs with `git`.